feat(project): change project structure to client-server
