+++ Cargo.toml
@@ -10,6 +10,4 @@ opt-level = 3
 
 [workspace.dependencies]
 bevy = { version = "0.13" }
-bevy_editor_pls = { version = "0.8" }
-leafwing-input-manager = { version = "0.13" }
 thiserror = { version = "1" }
