+++ tygydyk/src/bin/client.rs
@@ -1,0 +1,8 @@
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
+
+use bevy::prelude::App;
+use tygydyk::feature::plugin::TygydykClientPlugin;
+
+fn main() {
+    App::new().add_plugins(TygydykClientPlugin).run();
+}
