+++ tygydyk/src/bin/server.rs
@@ -1,0 +1,6 @@
+use bevy::prelude::App;
+use tygydyk::feature::plugin::TygydykServerPlugin;
+
+fn main() {
+    App::new().add_plugins(TygydykServerPlugin).run();
+}
