--- tygydyk/src/feature/camera/component.rs
@@ -1,4 +1,0 @@
-use bevy::prelude::{Component, Reflect};
-
-#[derive(Component, Debug, Reflect)]
-pub struct TygydykCamera;
