--- tygydyk/src/feature/camera/consts.rs
@@ -1,2 +1,0 @@
-pub const CAMERA_FAR: f32 = 1.;
-pub const CAMERA_NEAR: f32 = -1.;
