--- tygydyk/src/feature/camera/mod.rs
@@ -1,10 +1,0 @@
-mod bundle;
-mod component;
-mod consts;
-mod plugin;
-mod system;
-
-pub use bundle::*;
-pub use component::*;
-pub use consts::*;
-pub use plugin::*;
