--- tygydyk/src/feature/input/component.rs
@@ -1,1 +1,0 @@
-
