--- tygydyk/src/feature/input/mod.rs
@@ -1,6 +1,0 @@
-mod component;
-mod plugin;
-mod resource;
-
-pub use plugin::*;
-pub use resource::*;
