--- tygydyk/src/feature/kitten/mod.rs
@@ -1,8 +1,0 @@
-mod component;
-mod consts;
-mod plugin;
-mod system;
-
-pub use component::*;
-pub use consts::*;
-pub use plugin::*;
