+++ tygydyk/src/feature/mod.rs
@@ -1,8 +1,1 @@
-mod camera;
-mod input;
-mod kitten;
-mod movement;
-mod platform;
-mod plugin;
-
-pub use plugin::TygydykPlugin;
+pub mod plugin;
