--- tygydyk/src/feature/movement/mod.rs
@@ -1,6 +1,0 @@
-mod component;
-mod plugin;
-mod system;
-
-pub use component::*;
-pub use plugin::*;
