+++ tygydyk/src/feature/plugin/mod.rs
@@ -1,0 +1,5 @@
+mod client;
+mod server;
+
+pub use client::TygydykClientPlugin;
+pub use server::TygydykServerPlugin;
